requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
neomodel==5.3.0
python-dotenv==1.0.0
pydantic==2.5.0
setuptools>=68.0.0
email-validator>=2.0.0  # Add this line